+++ build.rs
@@ -1,5 +1,5 @@
-use std::error::Error;
 use std::env;
+use std::error::Error;
 use std::path::Path;
 
 fn main() -> Result<(), Box<dyn Error>> {
@@ -10,7 +10,7 @@ fn main() -> Result<(), Box<dyn Error>> {
     let out_dir = env::var("OUT_DIR")?;
     let out_path = Path::new(&out_dir);
     println!("cargo:warning=OUT_DIR is: {}", out_path.display());
-    
+
     // Ensure we're not creating any files in the source directory
     // If needed, only write files to out_path.join("some_file")